Implements local cache. Now fix dependency problem with rabbit logger

+++ file_storage/src/file_storage_server.py
@@ -1,6 +1,5 @@
 import os
 import uuid
-from typing import Dict
 
 import uvicorn
 from fastapi import FastAPI, File, UploadFile, HTTPException
@@ -9,23 +8,33 @@ from fastapi.responses import FileResponse
 from DicomFlowLib.log import CollectiveLogger
 
 
-class FileStorage(FastAPI):
+class FileStorageServer(FastAPI):
     def __init__(self, logger: CollectiveLogger,
                  base_dir: str,
-                 file_storage_host: str,
-                 file_storage_port: int,
-                 suffix: str = ".tar"):
+                 host: str,
+                 port: int,
+                 suffix: str = ".tar",
+                 allow_post: bool = True,
+                 allow_get: bool = True,
+                 allow_clone: bool = True,
+                 allow_delete: bool = True):
         super().__init__()
 
         self.base_dir = base_dir
         os.makedirs(self.base_dir, exist_ok=True)
         self.logger = logger
         self.suffix = suffix
-        self.file_storage_host = file_storage_host
-        self.file_storage_port = file_storage_port
+        self.host = host
+        self.port = port
+        self.allow_post = allow_post
+        self.allow_get = allow_get
+        self.allow_delete = allow_delete
+        self.allow_clone = allow_clone
 
         @self.post("/")
         def post(tar_file: UploadFile = File(...)):
+            if not self.allow_post:
+                raise HTTPException(status_code=405, detail="Method not allowed")
             uid = str(uuid.uuid4())
             self.logger.debug(f"Putting file on uid: {uid}", finished=False)
 
@@ -41,6 +50,8 @@ class FileStorage(FastAPI):
 
         @self.put("/")
         def clone(uid: str):
+            if not self.allow_clone:
+                raise HTTPException(status_code=405, detail="Method not allowed")
             new_uid = str(uuid.uuid4())
             self.logger.debug(f"Clone file on uid: {uid} to new uid: {new_uid}", finished=False)
 
@@ -52,7 +63,11 @@ class FileStorage(FastAPI):
 
         @self.get("/")
         def get(uid: str) -> FileResponse:
+            if not self.allow_get:
+                raise HTTPException(status_code=405, detail="Method not allowed")
+
             self.logger.debug(f"Serving file with uid: {uid}", finished=False)
+
             assert self.file_exists(uid)
 
             self.logger.debug(f"Serving file with uid: {uid}", finished=True)
@@ -60,6 +75,8 @@ class FileStorage(FastAPI):
 
         @self.delete("/")
         def delete(uid: str):
+            if not self.allow_delete:
+                raise HTTPException(status_code=405, detail="Method not allowed")
             self.logger.debug(f"Deleting file with uid: {uid}", finished=False)
             assert self.file_exists(uid)
             os.remove(self.get_file_path(uid))
@@ -76,4 +93,4 @@ class FileStorage(FastAPI):
         return b
 
     def start(self):
-        uvicorn.run(app=self, host=self.file_storage_host, port=self.file_storage_port)
+        uvicorn.run(app=self, host=self.host, port=self.port)
